update index file refactor routing
