+++ react-node-video-streaming/server/.gitignore
@@ -23,3 +23,6 @@ yarn-debug.log*
 yarn-error.log*
 
 .env
+*lock*
+tmp/
+uploads/
