--- react-node-video-streaming/server/app.js
@@ -1,77 +1,0 @@
-require('dotenv').config();
-const express = require('express');
-const mongoose = require("mongoose");
-
-const fs = require('fs');
-const path = require('path');
-const cors = require('cors');
-const {PORT} = require("./constants/constants");
-const {MONGO_URI} = require("./constants/constants");
-const bodyParser = require('body-parser');
-const io = require('socket.io')(3000)
-
-
-// connect to mongodb
-mongoose.connect(MONGO_URI, {
-    useUnifiedTopology: true,
-    useNewUrlParser: true,
-})
-    .then(res => {
-    console.log("connected to mongodb");
-})
-    .catch(err => {
-        console.log(err);
-    })
-
-
-const app = express();
-
-// use middleware to enable cors
-app.use(cors());
-app.use(bodyParser.json({limit: "50mb"}));
-app.use(bodyParser.urlencoded({limit: "5gb", extended: true, parameterLimit:50000}));
-
-
-app.get('/video', (req, res) => {
-    const path = `assets/frag_bunny.mp4`;
-    const stat = fs.statSync(path);
-    const fileSize = stat.size;
-    const range = req.headers.range;
-    console.log(range);
-    if (range) {
-        const parts = range.replace(/bytes=/, "").split("-");
-        const start = parseInt(parts[0], 10);
-        const end = parts[1] && parseInt(parts[1], 10) < fileSize
-            ? parseInt(parts[1], 10)
-            : fileSize-1;
-        const chunksize = (end-start) + 1;
-        const file = fs.createReadStream(path, {start, end});
-        const head = {
-            'Access-Control-Expose-Headers': 'Content-Range',
-            'Content-Range': `bytes ${start}-${end}/${fileSize}`,
-            'Accept-Ranges': 'bytes',
-            'Content-Length': chunksize,
-            'Content-Type': 'video/mp4',
-        };
-        res.writeHead(206, head);
-        file.pipe(res);
-    } else {
-        const head = {
-            'Content-Length': fileSize,
-            'Content-Type': 'video/mp4',
-        };
-        res.writeHead(200, head);
-        fs.createReadStream(path).pipe(res);
-    }
-});
-
-app.listen(PORT, () => {
-    console.log('Listening on port 8000!')
-});
-
-// Socket.io realtime
-io.on('connection', (socket) => {
-    socket.on('disconnect', () => {
-        console.log('user disconnected');
-    });
-});
