+++ react-node-video-streaming/server/package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "dev": "nodemon app.js",
+    "dev": "nodemon index.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "",
@@ -14,9 +14,15 @@
     "cors": "^2.8.5",
     "dotenv": "^10.0.0",
     "express": "^4.17.1",
+    "fluent-ffmpeg": "^2.1.2",
+    "fs-extra": "^10.0.0",
+    "googleapis": "^39.2.0",
+    "mime-types": "^2.1.34",
     "mongodb": "^4.2.0",
     "mongoose": "^6.0.13",
+    "multer": "^1.4.3",
     "nodemon": "^2.0.15",
-    "socket.io": "^4.3.2"
+    "socket.io": "^4.3.2",
+    "uuid": "^8.3.2"
   }
 }
